requirements.txt
# Core dependencies
anthropic>=0.18.0
ollama>=0.1.7
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Testing and evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.2.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
jinja2>=3.1.0

# Utilities
pyyaml>=6.0
colorama>=0.4.6
rich>=13.7.0
click>=8.1.0

# Code analysis
ast-comments>=1.2.0
black>=23.0.0
autopep8>=2.0.0

# Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# JSON handling
jsonschema>=4.20.0